requirements.txt
numpy==2.1.3
Pillow==11.1.0
streamlit==1.43.2
pytest==8.3.5
